Refactor search scripts so that navBar search functions across all pages.
