+++ public/js/navbar-search.js
@@ -1,79 +1,4 @@
 import { handleErrors, api, addMarkers } from './utils.js';
-window.initMap = (targets) => {
-	const sf = { lat: 37.773, lng: -122.431 };
-	// const test2 = { lat: 37.777, lng: -122.418 };
-	// const test3 = { lat: 37.779, lng: -122.421 };
-	// const test4 = { lat: 37.772, lng: -122.414 };
-
-	//icons
-	const blueFlagIcon = 'http://maps.google.com/mapfiles/ms/icons/flag.png';
-	const purpleIcon = 'http://maps.google.com/mapfiles/ms/icons/purple.png';
-	const purpleIconDot = 'http://maps.google.com/mapfiles/ms/icons/purple-dot.png';
-
-
-	const options = {
-		center: sf,
-		zoom: 13
-	};
-	const map = new google.maps.Map(document.getElementById('map'),
-		options);
-
-
-	// const markers = [
-	// 	{ coords: sf, iconImage: purpleIcon, content: '<h3>Resto Info' },
-	// { coords: test2, iconImage: purpleIconDot, content: '<h3>Resto Info' },
-	// { coords: test3, iconImage: blueFlagIcon, content: '<h3>Resto Info' },
-	//];
-	console.log("targets: ", targets)
-	if (targets) {
-		targets.forEach((target) => {
-			addMarker({ coords: { lat: parseFloat(target.lat), lng: parseFloat(target.lon) }, iconImage: purpleIconDot, content: `<h5>${target.name}</h5> <h6> - ${target.averageRating} Stars</h6>` });
-		});
-	}
-	// addMarker({
-	// 	coords: sf,
-	// 	iconImage: 'http://maps.google.com/mapfiles/ms/icons/flag.png',
-	// 	content: '<h3>Resto Info'
-	// });
-	// addMarker({
-	// 	coords: test2
-	// });
-	// addMarker({ postion: sf, map: map });
-
-	// const marker = new google.maps.Marker({
-	// 	position: sf,
-	// 	map: map
-	// });
-	// const infoWindow = new google.maps.InfoWindow({
-	// 	content: '<h1>Test Info</h1>'
-	// });
-	// marker.addListener('click', () => {
-	// 	infoWindow.open(map, marker);
-	// })
-
-	function addMarker(props) {
-		const marker = new google.maps.Marker({
-			position: props.coords,
-			map: map,
-		});
-
-		if (props.iconImage) {
-			marker.setIcon(props.iconImage);
-		}
-		//check content
-		if (props.content) {
-			const infoWindow = new google.maps.InfoWindow({
-				content: props.content
-			});
-			marker.addListener('mouseover', () => {
-				infoWindow.open(map, marker);
-			});
-			marker.addListener('mouseout', () => {
-				infoWindow.close(map, marker);
-			})
-		}
-	}
-}
 
 document.addEventListener('DOMContentLoaded', (e) => {
 	// NavBar Selectors
@@ -88,97 +13,7 @@ document.addEventListener('DOMContentLoaded', (e) => {
 	const searchField = document.querySelector('.searchField');
 	const collapseButton = document.querySelector('.navbar-collapse');
 
-	// const url = window.location.href;
 
-	function generateCards(businesses) {
-		const businessCardsHTML = businesses.map((business) => {
-			return `
-				<a class="business-cards" href="/businesses/${business.id}">
-					<div class="card mt-2 stars-business" id="business-${business.id}">
-						<div class="card-body">
-							<div class="card-text-name">${business.name}</div>
-							<div class="stars">
-								<span class="business-star star-${business.id}"></span>
-								<span class="business-star star-${business.id}"></span>
-								<span class="business-star star-${business.id}"></span>
-								<span class="business-star star-${business.id}"></span>
-								<span class="business-star star-${business.id}"></span>
-							</div>
-							<h3 class="card-text">${business.Tag.type}</h3>
-							<div class="card-text">${business.address}</div>
-						</div>
-					</div>
-				</a>
-				`;
-		});
-		businessCardContainer.innerHTML = businessCardsHTML.join('');
-	}
-	function findAverageRating(businesses) {
-		businesses.forEach((business) => {
-			let ratingsArray = [];
-			business.Reviews.forEach((rating) => {
-				ratingsArray.push(parseInt(rating.businessRating));
-			});
-			const sumRating = ratingsArray.reduce((a, b) => {
-				return a + b;
-			}, 0);
-			const averageRating = sumRating / ratingsArray.length;
-			business.averageRating = averageRating;
-			const businessStars = document.querySelectorAll(`.star-${business.id}`);
-			businessStars.forEach((star, index) => {
-				if (averageRating > index) {
-					star.classList.add('rated');
-				} else {
-					star.classList.remove('rated');
-				}
-			});
-			const businessId = document.getElementById(`business-${business.id}`);
-			businessId.setAttribute('data-rating', averageRating);
-		});
-	}
-
-	//add map data for search results
-	function addMapMarkers(businesses) {
-		addMarkers();
-	}
-
-	async function fetchBusinessSearch(body) {
-		try {
-			const res = await fetch(`${api}businesses/search`, {
-				method: 'POST',
-				body: JSON.stringify(body),
-				headers: {
-					'Content-Type': 'application/json'
-				}
-			});
-
-			const { businesses } = await res.json();
-			const test3 = { lat: 37.779, lng: -122.421 };
-
-			// declare ratings array to store the ratings
-			generateCards(businesses);
-			findAverageRating(businesses);
-			//addMapMarkers(businesses);
-			document.getElementById('map').classList.remove('map-default');
-			window.initMap(businesses);
-
-			//pin locations currently do not persist. Better to not persist search results to match functionality?
-			//dynamically adjust styling for map div height?
-		} catch (err) {
-			handleErrors(err);
-		}
-	}
-
-	// Business selectors
-	const businessCardContainer = document.querySelector('.business-card-container');
-
-	dropDown.addEventListener('click', () => {
-		document.querySelector('.dropdown-menu').classList.toggle('show');
-	});
-	collapseButton.addEventListener('click', () => {
-		// TODO: show collapsed elements when clicking button
-		collapseButton.classList.toggle('show');
-	});
 	dropDownMenu.addEventListener('click', (event) => {
 		// help me dry up this code please
 		console.log(event.target);
@@ -194,31 +29,12 @@ document.addEventListener('DOMContentLoaded', (e) => {
 		}
 	});
 
-	// check to see if there is already a search value in the session storage
-	const sessionSearchValue = sessionStorage.getItem('SEARCH_VALUE');
-	if (sessionSearchValue) {
-		document.getElementById('navbarSearch').value = sessionSearchValue;
-		const body = {
-			name: sessionSearchValue
-		};
-		fetchBusinessSearch(body);
-		//window.location.href = '/search';
-	}
-
 	// set up the event listener for the submit button
 	navbarSearchForm.addEventListener('submit', async (e) => {
 		// Prevent the default behavior of the submit button
 		e.preventDefault();
-		//TODO: when implementing search functionality
-
-
 		const searchValue = document.getElementById('navbarSearch').value;
-		console.log("Form activataed!", searchValue)
 		sessionStorage.setItem('SEARCH_VALUE', searchValue);
-		const body = {
-			name: searchValue
-		};
-		fetchBusinessSearch(body);
-		//window.location.href = '/';
+		window.location.href = '/search';
 	});
 });
