+++ views/search.pug
@@ -4,6 +4,6 @@ block content
   //include google-map.pug
   h1 Search Results
   .business-card-container
-  script(type="module" src="/js/navbar-search.js")
-  script(src=`https://maps.googleapis.com/maps/api/js?key=${mapAPI}&callback=initMap` async='' defer='')
+  script(type="module" src="/js/search.js")
+  script(src=`https://maps.googleapis.com/maps/api/js?key=${mapAPI}&callback=initMap`)
